Move message handling code into its own submodule
This is to help isolate code boundaries and make it easier to understand system behaviour.
